+++ packages/EngineBase/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mj-biz-apps/accounting-engine-base
 
+## 0.13.0
+
+### Patch Changes
+
+- @mj-biz-apps/accounting-entities@0.13.0
+
 ## 0.12.0
 
 ### Patch Changes